requirements.txt
cloudinary==1.44.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-storages==1.14.6
djangorestframework==3.16.1
pillow==11.3.0
python-decouple==3.8
python-dotenv==1.1.1
qrcode==8.2
psycopg2==2.9.10
whitenoise==6.9.0
gunicorn==23.0.0
requests==2.32.5
django-jsonfield-backport==1.0.5 
django-user-agents==0.4.0